+++ apps/website/src/app/components/AppBreadcrumbs.tsx
@@ -1,0 +1,5 @@
+import { Breadcrumbs } from '@mui/material';
+
+export function AppBreadcrumbs() {
+    return <Breadcrumbs></Breadcrumbs>;
+}
