+++ apps/website/src/app/components/Logo.tsx
@@ -1,15 +1,15 @@
 import { Box } from '@mui/material';
 
-import img from '../../../assets/Logo.png';
-import { urls } from '../../util/routes';
+import img from '../../assets/Logo.png';
+import { urls } from '../util/routes';
 
 export function Logo() {
     return (
         <Box height="4rem" width="7.5rem">
-            <a>
+            <a href={urls.home}>
                 <img
                     src={img}
-                    alt="applept16 Logo"
+                    alt="Ambrosia Logo"
                     height="200"
                     width="200"
                     style={{
