+++ apps/website/src/app/pages/login/LoginPage.tsx
@@ -10,11 +10,11 @@ import {
     useIsLoggedIn,
 } from '../../elf/self-user/SelfUser.repository';
 import { navTo, urls } from '../../util/routes';
-import { AppTypography } from '../common/AppTypography';
-import { AppButton } from '../common/button/AppButton';
-import { AppForm } from '../common/form/AppForm';
-import { AppInput } from '../common/form/AppInput';
-import { Page } from '../common/Page';
+import { AppTypography } from '../../components/AppTypography';
+import { AppButton } from '../../components/button/AppButton';
+import { AppForm } from '../../components/form/AppForm';
+import { AppInput } from '../../components/form/AppInput';
+import { Page } from '../../components/Page';
 
 export function LoginPage() {
     const isLoggedIn = useIsLoggedIn();
